+++ Backer/Talk/source/PLA.ts
@@ -1,0 +1,7 @@
+import * as Truth from "truth-compiler";
+
+export interface PLABase
+{
+  typePath: string[];
+}
+
