+++ Backer/Talk/source/System.ts
@@ -1,5 +1,4 @@
 import { Query } from "./Query";
-import * as Operations from "./Operations";
 import * as Truth from "truth-compiler";
 
 /**
@@ -56,28 +55,3 @@ export class System
 		return new Query(this.doc.types.slice());
 	}
 }
-
-async function main() 
-{
-	await System.fromFile("./example.truth");
-	const doc = System.this.doc;
-	const types = doc.types;
-
-	const getType = (name: string): Truth.Type => 
-	{
-		const type = types.filter(t => t.name === name)[0];
-		if (!type) throw new Error();
-		return type;
-	};
-	const Employee = getType("Employee");
-
-	// #- Query...
-
-	const query = System.this.query();
-	query.addOperation(new Operations.IsOperation(Employee));
-	const ret = query.run();
-
-	debugger;
-}
-
-main();
