+++ Backer/Talk/source/Talk.ts
@@ -1,0 +1,5 @@
+export function tt() {}
+
+export namespace tt {
+	export function is() {}
+}
